chore(ground-zero): Initiate fresh start for 'protonup-ng' project
This commit signifies the initiation of a fresh start for the
'protonup-ng' project. All existing content is being removed to pave
the way for a comprehensive rewrite. The new development will take
place in the 'ground-zero' branch.

While 'protonup-ng' is a hardfork of 'protonup', it's important to note
that the original project's code will not be reused. Instead, we're
taking inspiration from the original idea and building upon it in our
unique way.